split datapoint_source module;
