+++ core/src/datapoint_source/custom_ext_script.rs
@@ -1,0 +1,37 @@
+use derive_more::From;
+use thiserror::Error;
+
+use super::DataPointSource;
+use super::DataPointSourceError;
+
+#[derive(Debug, From, Error)]
+pub enum ExternalScriptError {
+    #[error("external script child process error: {0}")]
+    ChildProcess(std::io::Error),
+    #[error("String from bytes error: {0}")]
+    StringFromBytes(std::string::FromUtf8Error),
+    #[error("Parse i64 from string error: {0}")]
+    ParseInt(std::num::ParseIntError),
+}
+
+#[derive(Debug, Clone)]
+pub struct ExternalScript(String);
+
+impl ExternalScript {
+    pub fn new(script_name: String) -> Self {
+        ExternalScript(script_name)
+    }
+}
+
+impl DataPointSource for ExternalScript {
+    fn get_datapoint(&self) -> Result<i64, DataPointSourceError> {
+        let script_output = std::process::Command::new(&self.0)
+            .output()
+            .map_err(ExternalScriptError::from)?;
+        let datapoint_str =
+            String::from_utf8(script_output.stdout).map_err(ExternalScriptError::from)?;
+        datapoint_str
+            .parse()
+            .map_err(|e| DataPointSourceError::from(ExternalScriptError::from(e)))
+    }
+}
